Unset $body to make sure GET request get nor marked as bad requests
On ticket creation there is first a search for open tickets which is a POST request. Then the module does makes a GET request to get all custom fields from jira. Unfortunatly a GET request with a body will be a bad request. This patch makes sure the $body will be removed from the reused session if there is no body needed for the request.